added Post.get_boolean_fields
The posts_controller had a hardcoded method
to convert "1" parameters to true if they corresponded
to post fields. This was
	a) out of date
	b) non-dynamic

Post.get_boolean_fields queries the MongoID list of
fields, and returns the names of the ones that are boolean.

This is, in turn, used by the PostsController
( and anything else that may come along )