Ported admin UI from Obtvse
This incorporates MOST of the functionality
from the new admin UI from Obtvse

This does not include the auto-save functionality.
While this may be added in the future it was adding
complications, and it seemed best to simply get the
improved Admin UI working and out to people.

Many of the files required to support it ARE
present and just need to be tweaked to get it all
working corectly in the new environment.