now storing content type of BinaryFiles
with uploaded files it will get it directly
from the content_type data passed to the server.

with linked files (non s3) it will make an attempt
at extrapolating it from the extension. When there's
no extension or the extension is unrecognized it will
just default to 'application/octet-stream'

which maynot be the best call...

The content types it is now storing are not
currently used in the app, but it seemed a good thing
to store for usage down the road.