Added a markdown feed
Since Markdown is predominately used by geeks at the
moment, it stands to reason that we might have
automated systems sucking down feeds, and it's
a hell of a lot easier to parse markdown than HTML,

So why not have a markdown feed

/md_feed will now serve a standard Atom feed but with the
content type set to "text". Additionally I've added a
"markup='markdown'" attribute, which may, or may not
be helpful.